fixed bugs to save and load double from xml

+++ BalanceMenu.cpp
@@ -17,11 +17,11 @@ void BalanceMenu::addExpense()
 void BalanceMenu::showBalanceCurrentMonth()
 {
     system("cls");
-    int balance = 0;
+    double balance = 0;
     cout << "Przychody w obecnym miesiacu: "<< endl << endl;
-    int incomesSum = income.showBalanceCurrentMonth();
+    double incomesSum = income.showBalanceCurrentMonth();
     cout << endl << "Wydatki w obecnym miesiacu: "<< endl;
-    int expensesSum = expense.showBalanceCurrentMonth();
+    double expensesSum = expense.showBalanceCurrentMonth();
     cout << endl << "Suma przychodow w obecnym miesiacu wynosi " << incomesSum << endl;
     cout << endl << "Suma wydatkow w obecnym miesiacu wynosi " << expensesSum << endl;
     balance = incomesSum + expensesSum;
@@ -31,11 +31,11 @@ void BalanceMenu::showBalanceCurrentMonth()
 void BalanceMenu::showBalancePreviousMonth()
 {
     system("cls");
-    int balance = 0;
+    double balance = 0;
     cout << "Przychody w poprzednim miesiacu: "<< endl << endl;
-    int incomesSum = income.showBalancePreviousMonth();
+    double incomesSum = income.showBalancePreviousMonth();
     cout << "Wydatki w poprzednim mesiacu: "<< endl;
-    int expensesSum = expense.showBalancePreviousMonth();
+    double expensesSum = expense.showBalancePreviousMonth();
     cout << endl << "Suma przychodow w poprzednim miesiacu wynosi " << incomesSum << endl;
     cout << endl << "Suma wydatkow w poprzednim miesiacu wynosi " << expensesSum << endl;
     balance = incomesSum + expensesSum;
@@ -54,11 +54,11 @@ void BalanceMenu::showBalanceSelectedPeriod()
     cout << " do ";
     endingDate = AuxiliaryMethods::loadLine();
     if(DateOperation::checkSelectedDates(startingDate,endingDate)){
-        int balance = 0;
+        double balance = 0;
         cout << "Przychody w wybranym okresie: "<< endl << endl;
-        int incomesSum = income.showBalanceSelectedPeriod(startingDate,endingDate);
+        double incomesSum = income.showBalanceSelectedPeriod(startingDate,endingDate);
         cout << "Wydatki w wybranym okresie: "<< endl;
-        int expensesSum = expense.showBalanceSelectedPeriod(startingDate,endingDate);
+        double expensesSum = expense.showBalanceSelectedPeriod(startingDate,endingDate);
         cout << endl << "Suma przychodow w wybranym okresie wynosi " << incomesSum << endl;
         cout << endl << "Suma wydatkow w wybranym okresie wynosi " << expensesSum << endl;
         balance = incomesSum + expensesSum;
